requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.43.1
accelerate>=0.25.0
pandas>=1.5.0
numpy>=1.24.0
h5py>=3.8.0
Pillow>=9.0.0
tqdm>=4.65.0
wandb>=0.16.0
